Add support for WellRNG 512a
This is a "new" random number generator. Derived from code published at:

http://www.iro.umontreal.ca/~simardr/rng/WELL512a.c

See also:

http://www.iro.umontreal.ca/~panneton/WELLRNG.html
http://www.iro.umontreal.ca/~panneton/well/WELL512a.c

lfsr04.pdf: F. Panneton, P. L'Ecuyer and M. Matsumoto, "Improved
Long-Period Generators Based on Linear Recurrences Modulo 2"
http://www.iro.umontreal.ca/~lecuyer/myftp/papers/lfsr04.pdf

<QUOTE>
In this paper, we propose new generators, with better equidistribution and “bit-mixing”
properties for equivalent period length and speed. Approximately half of the coefficients
of the characteristic polynomial of these generators are nonzero. The state of our new
generators evolves in a more chaotic way than for the Mersenne twister.
</QUOTE>

Many thanks to Pierre L'Ecuyer for GPL'ing the code so that it could
be used in this project.