Look up RT authors in the DATA section aliases as well as AUTHORS
This patch is a little gross but if fixes a bug where an author wasn't found
at all. The author in question was in the git log as "bulk88 (via RT)
<perlbug-followup@perl.org>". Note the lack of a name before the "(via RT)"
part.

This whole thing is a bit of a mess. I think this whole script could be
re-organized in a simpler way, perhaps.