regcomp.c: Remove unnecessary flag setting
The function cl_anything() sets things up so that the synthetic start
class parameter will match any character.  But this flag doesn't
contribute to that, as it sets things to match under certain conditions
only, and the characters it matches already match by the other actions
of the routine (the ANYOF_BITMAP_SETALL just above).