regcomp.c: Clean up ANYOF_CLASS handling.
The ANYOF_CLASS flag is used in ANYOF nodes (for [bracketed] and the
synthetic start class) only when matching something like \w, [:punct:]
etc., under /l (locale).  It should not be set unless /l is specified.
However, it was always getting set for the synthetic start class.  This
commit fixes that.  The previous code was masking errors in which it was
being tested for unnecessarily, and for much of the 5.17 series, the
synthetic start class was always set to test for locale, which was a
waste of cpu when no locale was specified.