Add new regnode for synthetic start class
This creates a regnode specifically for the synthetic start class, which
is a type of ANYOF node.  The flag bit previously used to denote this is
removed.  This paves the way for this bit to be freed up, but first the
other use of this bit must also be removed, which will be done in the
next commit.

There are now three ANYOF-type regnodes.  This one should be called only
in one place in regexec.c.  The other special one is ANYOF_WARN_SUPER.
A synthetic start class node should not do any warning, so there is no
issue of having something need to be both types.