regcomp.c: Free up ANYOF flag bit
This frees up a flag bit for ANYOF regnodes.  The freed bit is currently
not needed for other uses; I decided to make the change now, while how
to do it was fresh in my mind.  There are fewer shifts and masks as a
result, as well.

This commit moves the information this bit contains to the otherwise
unused 'next_off' field in the synthetic start class.  This paradigm
could be used to pass information to the regex matching code for just
the synthetic start class, but the current bit is used just during
compilation.