locale.t: Add TODO tests
It turns out that Perl has always assumed that the Posix character
classes are closed under folding.  For example, if a character is in
[:alpha:], its fold will be in [:alpha:] as well.

This seems like a reasonable assumption except for two classes, where it
is almost certainly wrong.  If a character matches [:upper:], its fold
likely won't.  Same for [:lower:].  What this means is that a regex of
the form

    /[[:lower:]]/i

has never properly matched the uppercased versions of the characters in
the target string.

This commit adds TODO tests for these.