handy.h, regcomp.h, regexec.c: Sort initializers, switch()
Until recently, these were needed to be (or it made sense to be) in
numerical value of what the rhs of each #define evaluates to.  But now,
they are all initialized to something else, and the numerical value is
not even apparent.  Alphabetical order gives a logical ordering to help
a reader find things.