regcomp.c: Use data structure to remove special handling
[:digit:] matches nothing in the upper Latin1 range.  That means the
inversion list for the whole of Latin1 is the same as the one for the
ASCII range.  By copying the ASCII one into the Latin1's slot, we
eliminate the need for code to handle this case specially.

Future commits will remove the switch statement affected by this
commit entirely, and eliminating this special case becomes more
important given this direction.