regcomp.c: Fix upper loop limit
The upper limit test for this loop should be strictly less than, like
all other instances in the code for similar loops.  The class at
ANYOF_MAX is \v, which is not subject to locale rules, and this loop is
testing for which locale rule classes have been set.  Thus the loop
should not include \v.

Thanks to George Greer for spotting this.