Don't usenm on haiku
This cam up from building perl with nm enabled and make test_harness
croaking with the complaint that times wasn't implemented.

nm /system/lib/libroot.so | grep times produces:

00095540 T __times
00095460 T __times_beos
0009b690 t itimerspec_to_bigtimes__FRC10itimerspecRxT1
00095540 T times@@LIBROOT_1_ALPHA4
00095460 T times@LIBROOT_BASE
000981c8 t timesub
000958b4 T utimes

of which the two times@... entries appear to be version symbols.

The system perl on Haiku-OS alpha4.1 is built with -Uusenm