charnames: :alias alone implies :full
The documentation says this is how it should behave, but only 1 of the
three paths in the code did it, and in fact there was a test to the
contrary.

I'm only adding a test for one of the two fixed paths, as the other one
appears to require a weird file name.