Better error pos for grok_bslash_[xo]
These functions advance the parse pointer for the caller.  The regex
code has the infrastructure to output a marker as to where the error
was.  This commit simply moves the parse pointer past all the legal
digits in the input, which are likely supposed to be part of the number,
which makes it likely that the missing right brace point is just past
those.