Create new warnings category experimental::regex_sets
This on-by-default warning will be raised upon use of the
soon-to-be-delivered regex sets feature.  The syntax used is currently
invalid, so no feature activation is necessary.

The purpose of this warning is to make sure that any user is properly
notified that this feature is experimental and subject to change