grok_bslash_[ox]: Add param to silence non-portable warnings
If a hex or octal number is too big to fit in a 32 bit word, grok_oct
and grok_hex by default output a warning that it is a non-portable
value.  This new parameter to the grok_bslash functions can cause them
to shut up those warnings.  This is currently unused, but will be needed
in future commits.