add a dependency for dynamic extensions on $(LIBPERL)
This is necessary on some platforms (Haiku-OS, Cygwin and possibly
others) since EU::MM adds libperl.so/libperl.a (with their appropriate
extensions) as a dependency to the link step.

Added as a dependency on all platforms because it causes no practical
harm (fewer choices for parallel making) on the platforms where it isn't
needed.