Add back-compat (?[ ]) tests
This adds testing of (?[ ]), using the same tests, t/re/re_tests<
as are used by many of the regular expression .t files.  Basically, it
converts the [bracketed] character classes in these tests to the new
syntax and verifies that they work there.

Some tests won't work in one or the other, and the capability to skip
depending on the .t is added