Simplify ExtUtils::Install::_can_write_dir on VMS.
This reverts the code (but not test) portions of 3d55b451d9544fb.

The old solution of using catdir rather than a combination of
catdir and catpath works with and without extended filespecs
enabled, so there's no reason to maintain two versions, and thus
no reason to have all the boilerplate feature checking code at
the beginning of the module.