Add warnings for "\08", /\017/
This was discussed in thread
http://perl.markmail.org/thread/avtzvtpzemvg2ki2
but I never got around to this portion of the consensus, until now.

I did a cpan grep
http://grep.cpan.me/?q=%28^|[^\\]%29\\[0-7]{1%2C2}[8-9]&page=1

and eyeballing the results, saw three cases where this warning might
show up; one of which was for EBCDIC.  The others looked to be false
positives, such as in .css files.