Remove EFS escapes in unixify without directory components.
When converting a VMS file specification containing caret escapes
for extended characters to Unix format, we need to remove those
escapes.  We were doing so for paths containing directory
components but we also need to do it for bare filenames.  I.e.,
this was working:

disk:[dir]foo^%bar --> /disk/dir/foo%bar

but this was not (until now):

foo^%bar --> foo%bar