makedef.pl: Don't export inline fcns
An inline function is static to the file it is defined in.  But they can
be part of the public API if they are in #included header files and
defined in embed.fnc.  Normally such functions are written to the export
list on platforms that have this, but since they are static, you get a
linkage error.  This commit suppresses the writing of inline functions