Standardize removal of escapes in unixify.
When converting VMS format file specification to Unix format, we
need to translate sequences escaped with a caret in the VMS format
name into an unescaped form in the resulting Unix name.  We have
a function to do just that, but weren't using it in three of the
most important places that need it.  Using the function gets us
handling more cases than the simple versions we had inline.