Help ExtUtils::Manifest deal with VMS extended filespecs.
For one thing, if we're cleaning up filenames, lets put that code
in the sub called 'clean_up_filename' rather than in manifind.  And
since the language of the MANIFEST is really Unix-format filespecs,
convert to those.  This becomes especially important in handling
extended filespecs because unixification also unescapes escaped
characters.