Make readdir on VMS only unixify when asked.
We've been making our home-grown readdir implementation convert
all its results to Unix format since a096370a74e8.  Since readdir
returns relative paths, the usual volume and directory syntax
differences that distinguish VMS and Unix paths don't really apply.

But what does apply is whether the escaping of extended characters
is passed through or removed, and converting to Unix syntax removes
those escapes.  In order to enable extended filename syntax, though,
we're really going to need the escapes.  For example, rmdir will
not be able to remove a directory foo.bar if it's referred to as
[.foo.bar] because a dot is the traditional directory delimiter.
The directory needs to be specified as [.foo^.bar] for rmdir to
delete it successfully.

So make readdir return filespecs in native format unless explicitly
requested to report all filenames in Unix format.