Make extended filename syntax the default on VMS.
This feature has been available for over a decade but is still not
the default for the CRTL.  It seems time to make it the default
for Perl.  It means that directory names can have dots in them,
and filenames can contain multiple dots as well as spaces and
generally characters from the top row of the keyboard while the
shift key is held down.

Filenames with native syntax must escape characters in the so-called
extended character set with a caret ("^"), but filenames in Unix
syntax must not have the escapes.  Conversions between native and
Unix syntax will, respectively, add and remove these escapes.

Supporting Unicode in filenames depends on this feature but has
not yet been investigated.