[perl #116587] Data::Dumper intermittently fails tests on 5.8.*
In a nutshell, we don't care enough about 5.8 to try really hard to fix
(obscure) vstring problems there. But we care enough not to make DD fail
its tests because of such a corner case. We don't want to screw all
users that are trapped in the last decade.

Thus, skip certain tests pre-5.10.