utf8_heavy.pl: Add 'use re "/aa"'
This would speed up compilation and execution by a trivial amount, but
makes sure that no regular expression even tries to compile in Unicode,
which could cause recursive invocation of this.  Before this construct
was invented, special care had to be taken to keep that from happening;
those measures remain in effect, but this gives added insurance.