Remove the check for SVt_BIND from SvOK().
The original plan to use SVt_BIND to implement read-only aliases to
read-write values is unlikely to happen. More importantly, it's not going
to happen within a maint branch, so there's no reason to have the code
"just in case" it does. The code can easily be re-instated in blead if it
is needed in future. Nothing on CPAN is relying on this code. (Almost no
code on CPAN even references SVt_BIND.)

This effectively reverts part of commit 1cb9cd5016282146 from Dec 2006.