makedepend.SH: Deal with continuation lines
makedepend was failing to take into account the fact that preprocessor
directives, like any line, may be continued onto succeeding lines by a
trailing backslash.  This led to errors in the C preprocessing, which were
suppressed on all but os390 systems.o

This commit alleviates that, but it is not a complete solution.  It
deletes C comments, but this being sed, it doesn't catch all nuances of
C comments.  We only care, however, about comments on preprocessor
directive lines, and this is good enough to work on the current Perl
without generating any errors, on my system.

Spotted by John Goodyear.