Revert "Lookup overloaded assignment operators when trying to swap the arguments"
This reverts commit f041cf0f9c6469c41de8b73d5f7b426710c3ff8b.

This is reverted in reference to [perl #113834].  Although this commit
does fix a real bug, it introduces a new one that is arguably worse.
More importantly, it is a regression from 5.16.0.  We should produce a
test case that solves both bugs, then solve them both, then get that in
place.  Before 5.18.0 would be nice.  After 5.18.0 may have to do.