Loosen the passing requirement for a regex.
RT #116961: Olivier Menguéreported many test failures when installing
Data-Dumper v2.43 on pre-5.16.0 versions of Perl.  This patch addresses only
one of those issues by relaxing what is expected in a pattern match:
'42' vs. 42.  It doesn't matter for the test in question because what we are
testing is the "annotation" of array elements with their indices.