reorder intrpvar.h
Move more of the more commonly-used PL_ variables towards the front of the
file (and thus to the top of the interpreter struct on MULTIPLICITY
builds).

This helps ensure that "hot" variables are clustered together on the same
small number of cache lines, and also that the machine code to load them
will have shorter offsets, which on some architectures may be achieved
with shorter instructions.

The "hotness" has been determined purely by my subjective judgement rather
than any profiling. It's still open for the later to be done.

(Only simple shunting of whole lines has been done; no changes have been
made to individual lines.)