pp_entersub optimization
There is code in pp_entersub() that specifically handles the case where
AvARRAY(@_) != AvALLOC(@_), and tries to be more efficient than just
realloc()ing AvALLOC().

However, @_'s that have been shifted or messed with are usually cleaned up
or abandoned by CLEAR_ARGARRAY/POPSUB anyway, so this code is (probably)
never called.  So remove it.
In a worst case, it would just mean that realloc() is called
unnecessarily, which would be slower, but still correct.